Add images to readme
Add images to readme and rename window.py to main.py
