+++ README.MD
@@ -6,5 +6,22 @@ A project using OpenCV to play games and create filters.
 pip install -r requirements.txt
 ```
 
+## Startup
+
+    Once you have ran the SOCap.exe file, or the main.py file, you should see a window as such:
+    ![Main Window](./images/screenshots/main_window.PNG?raw=true)
+
+    From here, you can use the 'Webcam' button to open your device's webcam and begin applying filters, or you may use the 'Game' button to open your device's webcam and start using your hands as controllers.
+
+    Additionally, you may use the 'Rebind' button to open this window:
+    ![Main Window](./images/screenshots/rebind.PNG?raw=true)
+    Once open, you may enter keys in the order of Index finger, Middle finger, Ring finger, Pinky finger, Index + Middle finger, and Ring + Pinky finger.
+
+## Usage
+    
+    As mentioned above, to apply filters and use your left hand as a controller, you simply touch your desired finger to your thumb. For example, if you wish to blur the background behind you, simply touch the tip of your Index finger to the tip of your thumb.
+
 ## Contributors
+
 [Ravnit Lotay](https://github.com/Ravnit202), [Niral Patel](https://github.com/niralpatel02), [Nabeel Dar](https://github.com/NabeelDar)
+[EOF]
